Prior-value context-adaptive source model
Builds 256 models. Stores previous pixel value at that location in the frame to access that pixel model.

Moved files from other encoding schemes to another folder in data. Screenshots comparing a method I tried where I built a model for each location in the window and compared it to building a model for each pixel intensity value in screenshots folder in data.